Refractor : repository 확장성 고려하여 코드 수정 - Interface(CalculatorRepository) implements 하는 방식으로 MapCalculatorRepository 구현
